make Battle derived from Channel
The chat for each battle now takes place in a channel.
Channels use for this purpose have names prefixed by __battle__
When a client joins/leaves a battle, he is auto-joined/left to this channel.
 (-> such channels can register with ChanServ and set a founder, topic, operators, mutes, bans, etc.) 
To speak in a battle, use SAY/SAYEX/SAYFROM into its channel.

-- Changes to protocol -- 
- JOINBATTLE (server) has an extra arg, containing the name of the channel associated to the battle
- removed SAYBATTLEPRIVATE SAYBATTLEPRIVATEEX 
- deprecated SAYBATTLE SAYBATTLEEX (-> use SAY/SAYEX)
- added !registerbattle and !mute/!unmute to ChanServ
- deprecated SETCHANNELKEY and ChanServ !lock/!unlock (-> use !setkey)
- fix #283: LEFT is reflected back to the user who leaves a channel 

Clients should send the 'u' compatibility flag to signal that they are prepared for the above.
Full backwards compatibility exists for clients lacking 'u'.